third_party/WebKit/LayoutTests/http/tests/devtools/console/console-log-syntax-error-expected.txt

CONSOLE ERROR: line 3: Uncaught SyntaxError: Unexpected token )
Tests that syntax errors are logged into console and doesn't cause browser crash.

syntax-error.js:3 Uncaught SyntaxError: Unexpected token )
